chore: move migration to own package
