+++ .github/workflows/01-build-outputs.yml
@@ -27,6 +27,12 @@ jobs:
           name: db-ui-components-build
           path: packages/components/build
 
+      - name: ⏬ Download migration build
+        uses: actions/download-artifact@v4
+        with:
+          name: db-ui-migration-build
+          path: packages/migration/build
+
       - name: ⏬ Download output
         uses: actions/download-artifact@v4
         with:
