+++ packages/migration/package.json
@@ -1,0 +1,41 @@
+{
+  "name": "@db-ui/migration",
+  "version": "0.0.0",
+  "type": "module",
+  "description": "Provides basic tokens and assets based on DB UX Design System Core.",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/db-ui/mono.git"
+  },
+  "license": "DB Design License",
+  "bin": {
+    "@db-ui/migration": "build/index.js"
+  },
+  "main": "build.js",
+  "files": [
+    "build"
+  ],
+  "scripts": {
+    "build": "node esbuild.js",
+    "copy-build": "npm-run-all copy-build:*",
+    "copy-build:build": "cpr build ../../build-outputs/migration/build -o",
+    "copy-build:package.json": "cpr package.json ../../build-outputs/migration/package.json -o",
+    "copy-build:readme": "cpr README.md ../../build-outputs/migration/README.md -o",
+    "test": "vitest run --config vitest.config.ts",
+    "test:migration": "tsx src/cli.ts migration --type v005_v006 --src=./scss --dryRun=true"
+  },
+  "dependencies": {
+    "commander": "^12.0.0",
+    "glob": "^11.0.0",
+    "replace-in-file": "^8.2.0"
+  },
+  "devDependencies": {
+    "cpr": "3.0.1",
+    "tsx": "^4.19.2",
+    "vitest": "^2.1.8"
+  },
+  "publishConfig": {
+    "registry": "https://registry.npmjs.org/",
+    "access": "public"
+  }
+}
