+++ packages/migration/src/cli.ts
