+++ packages/migration/src/migration/v0.0.5-v0.0.6.ts
