+++ packages/migration/src/types.ts
@@ -1,3 +1,5 @@
+import { v005_v006 } from './migration/v0.0.5-v0.0.6';
+
 export type ProgrammOptionsType = {
 	name: string;
 	short?: string;
@@ -10,6 +12,6 @@ export type ProgrammOptionsType = {
 
 export type OptionsType = {
 	src: string;
-	type: (string | 'colorQ32024' | 'iconQ32024')[];
+	type: (string | 'colorQ32024' | 'iconQ32024' | 'v005_v006')[];
 	dryRun?: string | boolean;
 };
