+++ packages/migration/test/icon-q32024/icon-q32024.spec.ts
@@ -1,11 +1,11 @@
 import { describe, expect, test } from 'vitest';
-import { migrate } from '../../migration';
-import { ReplaceResult } from 'replace-in-file';
+import { type ReplaceResult } from 'replace-in-file';
+import { migrate } from '../../src/migration';
 
 describe('icon-q32024', () => {
 	test('check if changes are detected', async () => {
 		const result = migrate({
-			src: './scripts/test/icon-q32024',
+			src: './test/icon-q32024',
 			type: ['iconQ32024'],
 			dryRun: true
 		});
