+++ packages/migration/test/v0.0.5-v0.0.6/has-changes.txt
