+++ packages/migration/vitest.config.ts
@@ -2,6 +2,6 @@ import { defineConfig } from 'vitest/config';
 
 export default defineConfig({
 	test: {
-		include: ['scripts/**/*.{test,spec}.?(c|m)[jt]s?(x)']
+		include: ['test/**/*.{test,spec}.?(c|m)[jt]s?(x)']
 	}
 });
