+++ scripts/package.json
@@ -4,10 +4,11 @@
   "type": "module",
   "description": "contains scripts for monorepo",
   "scripts": {
-    "build": "npm-run-all build:foundations build:components",
+    "build": "npm-run-all build:foundations build:components build:migration",
     "build-outputs": "npm-run-all -p build-outputs:*",
     "build-outputs:components": "npm run copy-output --workspace=@db-ui/components",
     "build-outputs:foundations": "npm run copy-build --workspace=@db-ui/foundations",
+    "build-outputs:migration": "npm run copy-build --workspace=@db-ui/migration",
     "build-outputs:ngx-components": "npm run build --workspace=@db-ui/ngx-components",
     "build-outputs:react-components": "npm run build --workspace=@db-ui/react-components",
     "build-outputs:stencil-components": "npm run build --workspace=@db-ui/web-components",
@@ -24,6 +25,7 @@
     "build-showcases:vue": "npm run build --workspace=vue-showcase",
     "build:components": "npm run build --workspace=@db-ui/components",
     "build:foundations": "npm run build --workspace=@db-ui/foundations",
+    "build:migration": "npm run build --workspace=@db-ui/migration",
     "dev": "tsx dev.js",
     "dev:angular-components": "npm run dev:angular --workspace=@db-ui/components",
     "dev:plain-html-components": "npm run dev:html --workspace=@db-ui/components",
@@ -60,7 +62,7 @@
     "start-showcase:vue": "npm run dev --workspace=vue-showcase",
     "start:foundations": "npm run start --workspace=@db-ui/foundations",
     "test": "npm-run-all -p test:*",
-    "test:foundations-migration": "npm run test --workspace=@db-ui/foundations",
+    "test:migration": "npm run test --workspace=@db-ui/migration",
     "update:icon-fonts": "npm-run-all generate:icon-fonts generate:icon-types"
   },
   "devDependencies": {
